requirements.txt
python-dotenv==1.0.0
slack_sdk==3.21.3